Local server

+++ processes/dags/pubmed_migration_local.py
@@ -2,7 +2,6 @@ from airflow import DAG
 from airflow.operators.bash_operator import BashOperator
 from airflow.operators.dummy_operator import DummyOperator
 from airflow.operators.python_operator import PythonOperator
-from airflow.hooks.S3_hook import S3Hook
 from airflow.contrib.hooks.ftp_hook import FTPHook
 from airflow.contrib.hooks.mongo_hook import MongoHook
 from datetime import datetime, timedelta
@@ -16,8 +15,6 @@ import glob
 import json
 import pubmed_parser as pp
 
-# Setting up boto3 hook to AWS S3
-s3_hook = S3Hook('my_conn_S3')
 # Setting up FTP hook to pubmed ftp server
 ftp_hook = FTPHook('pubmed_ftp')
 # Setting up MongoDB hook to mlab server
@@ -26,8 +23,10 @@ mongodb_hook = MongoHook('mongo_default')
 
 def extract_original_name(filepath: str) -> str:
     """Extracts original name from tar.gz file
+
     Args:
         filepath (str): path of file.
+
     Returns:
         str: The original name
     """
@@ -37,6 +36,7 @@ def extract_original_name(filepath: str) -> str:
 
 def create_dir(dirname: str) -> None:
     """Creates directory even if exist
+
     Args:
         dirname (str): name of directory.
     """
@@ -45,15 +45,18 @@ def create_dir(dirname: str) -> None:
 
 
 def delete_dir(dirname: str) -> None:
-    """Deletes directory even if exist
+    """Deletes directory even if it exists or not
+
     Args:
         dirname (str): name of directory.
     """
-    shutil.rmtree(dirname)
+    if os.path.exists(dirname) and os.path.isdir(dirname):
+        shutil.rmtree(dirname)
 
 
 def extract_file(filepath: str, dest_dir: str, filter_pattern: str = None) -> None:
     """Extracts tar.gz file.
+
     Args:
         filepath (str): path of file.
         dest_dir (str): destination directory.
@@ -71,6 +74,7 @@ def extract_file(filepath: str, dest_dir: str, filter_pattern: str = None) -> No
 
 def extract_file_case_reports(filepath: str, dest_dir: str) -> None:
     """Extracts only case reports from tar.gz file.
+
     Args:
         filepath (str): path of file.
         dest_dir (str): destination directory.
@@ -80,6 +84,7 @@ def extract_file_case_reports(filepath: str, dest_dir: str) -> None:
 
 def delete_file(filepath: str) -> None:
     """Delete file.
+
     Args:
         filepath (str): path of file.
     """
@@ -89,6 +94,7 @@ def delete_file(filepath: str) -> None:
 
 def make_tarfile(output_filepath: str, source_dir: str) -> None:
     """Make tar.gz file.
+
     Args:
         output_filepath (str): path of output file.
         dest_dir (str): source directory.
@@ -97,16 +103,9 @@ def make_tarfile(output_filepath: str, source_dir: str) -> None:
         tar.add(source_dir, arcname=os.path.basename(source_dir))
 
 
-def delete_temp() -> None:
-    """Deletes temporary directory that processes files
-    """
-    root_dir = '/usr/local/airflow'
-    temp_dir = root_dir + '/' + 'temp'
-    delete_dir(temp_dir)
-
-
 def pubmed_get_text(pubmed_paragraph: dict) -> str:
     """Extracts text from pubmed_paragraph
+
     Args:
         pubmed_paragraph (dist): dict with pubmed_paragraph info
     """
@@ -115,14 +114,15 @@ def pubmed_get_text(pubmed_paragraph: dict) -> str:
 
 def pubmed_get_authors(pubmed_xml: dict) -> list:
     """Extracts authors from pubmed_paragraph
+
     Args:
         pubmed_xml (dist): dict with pubmed_xml info
     """
     return [" ".join(a[0:-1]) for a in pubmed_xml["author_list"]]
 
-
 def build_case_report_json(xml_path: str) -> json:
     """Makes and returns a JSON object from pubmed XML files
+
     Args:
         xml_path (str): path to input XML file
     """
@@ -133,7 +133,7 @@ def build_case_report_json(xml_path: str) -> json:
 
     case_report = {
         "pmID": pubmed_xml["pmid"],
-        "title": pubmed_xml["full_title"]
+        "title": pubmed_xml["full_title"],
         "messages": [],
         "source_files": [],
         "modifications": [],
@@ -163,32 +163,28 @@ def build_case_report_json(xml_path: str) -> json:
 
 
 def extract_pubmed_data() -> None:
-    """Extracts case-reports from pubmed data and stores result on S3
+    """Extracts case-reports from pubmed data and stores result on local server
     """
+    # to test specific tar files
+    #pattern = 'non_comm_use.A-B.xml.tar.gz'
     pattern = "*.xml.tar.gz"
     ftp_path = '/pub/pmc/oa_bulk'
     root_dir = '/usr/local/airflow'
-    temp_dir = os.path.join(root_dir, 'temp')
-    bucket_name = 'supreme-acrobat-data'
+    pubmed_dir = os.path.join(root_dir, 'pubmed')
+    original_dir = os.path.join(pubmed_dir, 'original')
     prefix = 'case_reports/pubmed/original'
 
-    # deleting old entries in the bucket
-    dest_path = 'case_reports/pubmed/original/'
-    wildcard = 'case_reports/pubmed/original/*.*'
-    old_klist = s3_hook.list_keys(bucket_name, prefix=dest_path, delimiter='/')
-    if isinstance(old_klist, list):
-        old_kmatches = [k for k in old_klist if fnmatch.fnmatch(k, wildcard)]
-        if len(old_kmatches) > 0:
-            s3_hook.delete_objects(bucket_name, old_kmatches)
+    delete_dir(original_dir)
+    create_dir(pubmed_dir)
+    create_dir(original_dir)
 
     filenames = ftp_hook.list_directory(ftp_path)
     filenames = list(
         filter(lambda filename: fnmatch.fnmatch(filename, pattern), filenames))
 
     for filename in filenames:
-        create_dir(temp_dir)
         remote_path = os.path.join(ftp_path, filename)
-        local_path = os.path.join(temp_dir, filename)
+        local_path = os.path.join(original_dir, filename)
 
         ftp_hook.retrieve_file(remote_path, local_path)
         o_path = extract_original_name(local_path)
@@ -196,18 +192,15 @@ def extract_pubmed_data() -> None:
         extract_file_case_reports(local_path, o_path)
         delete_file(local_path)
         make_tarfile(local_path, o_path)
-        key = os.path.join(prefix, filename)
-        s3_hook.load_file(
-            local_path, key, bucket_name=bucket_name, replace=True)
-        delete_dir(temp_dir)
-
+        delete_dir(o_path)
 
 def extract_pubmed_data_failure_callback(context) -> None:
-    delete_temp()
+    pass
 
 
 def join_json_data(filenames: str, dest_path: str) -> None:
     """Make a json file consisting of multiple json data
+
     Args:
         filenames (str): names of input json files
         dest_path (str): directory for combined json output
@@ -222,74 +215,52 @@ def join_json_data(filenames: str, dest_path: str) -> None:
 
 
 def transform_pubmed_data() -> None:
-    """Downloads tarfile from S3, forms JSON files from contents, and uploads to S3
+    """Downloads forms JSON files from contents of tarfile
     """
     root_dir = '/usr/local/airflow'
-    temp_dir = os.path.join(root_dir, 'temp')
-    src_path = 'case_reports/pubmed/original/'
-    dest_path = 'case_reports/pubmed/json/'
-    src_bucket_name = 'supreme-acrobat-data'
-    dest_bucket_name = 'supreme-acrobat-data'
-    wildcard_key = 'case_reports/pubmed/original/*.*'
-    klist = s3_hook.list_keys(src_bucket_name, prefix=src_path, delimiter='/')
-    key_matches = [k for k in klist if fnmatch.fnmatch(k, wildcard_key)]
-    create_dir(temp_dir)
-    filecount = 0
+    pubmed_dir = os.path.join(root_dir, 'pubmed')
+    json_dir = os.path.join(pubmed_dir, 'json')
+    original_dir = os.path.join(pubmed_dir, 'original')
 
-    # deleting old entries in the JSON folder
-    wildcard = 'case_reports/pubmed/json/*.*'
-    old_klist = s3_hook.list_keys(dest_bucket_name, prefix=dest_path, delimiter='')
-    if isinstance(old_klist, list):
-        old_kmatches = [k for k in old_klist if fnmatch.fnmatch(k, wildcard)]
-        if len(old_kmatches) > 0:
-            s3_hook.delete_objects(dest_bucket_name, old_kmatches)
+    delete_dir(json_dir)
+    create_dir(pubmed_dir)
+    create_dir(json_dir)
 
-    for key in key_matches:
-        filecount += 1
-        basename = os.path.basename(key)
-        local_path = os.path.join(temp_dir, basename)
-        o_path = extract_original_name(local_path)
+    file_path = os.path.join(original_dir, "*.xml.tar.gz")
+    filenames = [f for f in glob.glob(file_path)]
+
+    for filename in filenames:
+        basename = os.path.basename(filename)
+        local_path = os.path.join(original_dir, basename)
+        dest_path = os.path.join(json_dir, basename)
+        o_path = extract_original_name(dest_path)
         o_path_basename = os.path.basename(o_path)
-
-        obj = s3_hook.get_key(key, src_bucket_name)
-        obj.download_file(local_path)
-        extract_file(local_path, temp_dir)
+        extract_file(local_path, json_dir)
         glob_path = os.path.join(o_path, "*", "*.nxml")
-
         filenames = [f for f in glob.glob(glob_path)]
-        json_path = os.path.join(temp_dir, 'temp' + str(filecount) + '.json')
+        json_path = os.path.join(json_dir, o_path_basename + '.json')
         join_json_data(filenames, json_path)
-
-        key = os.path.join(dest_path, os.path.basename(
-            o_path_basename + ".json"))
-        s3_hook.load_file(
-            json_path, key, bucket_name=dest_bucket_name, replace=True)
-
-    delete_dir(temp_dir)
-
+        delete_dir(o_path)
 
 def transform_pubmed_data_failure_callback(context) -> None:
-    delete_temp()
+    pass
 
 
 def update_mongo() -> None:
     """Updates MongoDB caseReports
     """
-    src_path = 'case_reports/pubmed/json/'
-    src_bucket_name = 'supreme-acrobat-data'
-    wildcard_key = 'case_reports/pubmed/json/*.*'
+    root_dir = '/usr/local/airflow'
+    pubmed_dir = os.path.join(root_dir, 'pubmed')
+    json_dir = os.path.join(pubmed_dir, 'json')
 
-    klist = s3_hook.list_keys(src_bucket_name, prefix=src_path, delimiter='/')
-    key_matches = [k for k in klist if fnmatch.fnmatch(k, wildcard_key)]
+    file_path = os.path.join(json_dir, "*.json")
+    filenames = [f for f in glob.glob(file_path)]
 
-    for key in key_matches:
+    for filename in filenames:
         docs = []
-        obj = s3_hook.get_key(key, src_bucket_name)
-        file_content = obj.get()['Body'].read().decode('utf-8')
-
-        for line in file_content.splitlines():
-            json_content = json.loads(line)
-            docs.append(json_content)
+        with open(filename) as f:
+            for line in f:
+                docs.append(json.loads(line))
 
         collection = 'caseReports'
         filter_docs = [{'pmID': doc['pmID']} for doc in docs]
@@ -303,7 +274,7 @@ default_args = {
 }
 
 dag = DAG(
-    'pubmed_migration_AWS',
+    'pubmed_migration_local',
     default_args=default_args,
     description='Pubmed Migration',
     schedule_interval='@once',
@@ -312,7 +283,6 @@ dag = DAG(
 extract_pubmed_data_task = PythonOperator(
     task_id='extract_pubmed_data',
     python_callable=extract_pubmed_data,
-    on_failure_callback=extract_pubmed_data_failure_callback,
     dag=dag,
 )
